third_party/blink/web_tests/platform/generic/external/wpt/css/css-contain/container-queries/svg-foreignobject-child-container-expected.txt

This is a testharness.js-based test.
FAIL #inner querying #container inside foreignObject assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
